backend/requirements.txt
flask==2.0.0
flask-cors==3.0.10
SQLAlchemy==1.4.17
Flask-SQLAlchemy==2.5.1
pymysql==1.0.2
requests==2.7.0
Pillow==8.2.0
flask-msearch==0.2.9.1
whoosh==2.7.4
blinker==1.4
tinify==1.5.1
opencv-python==4.5.2.52
gunicorn==20.1.0
gevent==21.1.2
pynacl==1.4.0
boto3==1.17.101
cloudinary==1.26.0